Move all files out of the People folder
