+++ ui/src/PeopleMover/ArchivedPersonDrawer/PersonCard/PersonCard.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../../Styles/Colors';
+@import '../../../Styles/Colors';
 
 .archivedPersonCard {
   background: $gray-5;
