+++ ui/src/PeopleMover/ArchivedPersonDrawer/PersonCard/PersonCard.test.tsx
@@ -23,7 +23,7 @@ import {ViewingDateState} from 'State/ViewingDateState';
 import {IsReadOnlyState} from 'State/IsReadOnlyState';
 import {ModalContents, ModalContentsState} from 'State/ModalContentsState';
 import {RecoilObserver} from 'Utils/RecoilObserver';
-import PersonForm from '../PersonForm/PersonForm';
+import PersonForm from 'Common/PersonForm/PersonForm';
 import {renderWithRecoil} from 'Utils/TestUtils';
 import {Person} from 'Types/Person';
