+++ ui/src/PeopleMover/FullPageTests/ArchivedProduct.test.tsx
@@ -16,9 +16,9 @@
  */
 
 import React from 'react';
-import ArchivedProduct from './ArchiveProductsDrawer/ArchivedProduct/ArchivedProduct';
-import TestUtils, {renderWithRecoil} from '../Utils/TestUtils';
-import TestData from '../Utils/TestData';
+import ArchivedProduct from '../ArchiveProductsDrawer/ArchivedProduct/ArchivedProduct';
+import TestUtils, {renderWithRecoil} from 'Utils/TestUtils';
+import TestData from 'Utils/TestData';
 import {fireEvent, screen} from '@testing-library/react';
 
 jest.mock('Services/Api/SpaceClient');
